+++ CHANGES.rst
@@ -6,6 +6,7 @@ Changelog
 development
 ===========
 - Add example for annotating phenology data within Grafana
+- Change license to AGPL-3.0
 
 
 2020-12-27 0.1.2