+++ CHANGES.rst
@@ -7,6 +7,7 @@ development
 ===========
 - Add example for annotating phenology data within Grafana
 - Change license to AGPL-3.0
+- Modernize dependency versions. Drop support for Python 3.6.
 
 
 2020-12-27 0.1.2