+++ CHANGES.rst
@@ -5,6 +5,10 @@ Changelog
 
 development
 ===========
+
+
+2022-01-22 0.2.0
+================
 - Add example for annotating phenology data within Grafana
 - Change license to AGPL-3.0
 - Modernize dependency versions. Drop support for Python 3.6.